+++ tests/tests_unit/test_contextualization/test_pnid_parsing.py
@@ -24,11 +24,35 @@ def mock_parse(rsps):
 
 
 @pytest.fixture
+def mock_detect(rsps):
+    response_body = {"jobId": 456, "status": "Queued"}
+    rsps.add(
+        rsps.POST,
+        PNIDAPI._get_base_url_with_base_path() + PNIDAPI._RESOURCE_PATH + "/extractpattern",
+        status=200,
+        json=response_body,
+    )
+    yield rsps
+
+
+@pytest.fixture
 def mock_status_ok(rsps):
     response_body = {"jobId": 123, "status": "Completed", "svgUrl": "x"}
     rsps.add(
         rsps.GET,
         re.compile(PNIDAPI._get_base_url_with_base_path() + PNIDAPI._RESOURCE_PATH + "/\\d+"),
+        status=200,
+        json=response_body,
+    )
+    yield rsps
+
+
+@pytest.fixture
+def mock_status_pattern_ok(rsps):
+    response_body = {"jobId": 456, "status": "Completed", "items": []}
+    rsps.add(
+        rsps.GET,
+        re.compile(PNIDAPI._get_base_url_with_base_path() + PNIDAPI._RESOURCE_PATH + "/extractpattern" + "/\\d+"),
         status=200,
         json=response_body,
     )
@@ -80,3 +104,25 @@ class TestPNIDParsing:
         with pytest.raises(ModelFailedException) as exc_info:
             job.result
         assert "ContextualizationJob 123 failed with error 'error message'" == str(exc_info.value)
+
+    def test_detect(self, mock_detect, mock_status_pattern_ok):
+        patterns = ["ab{1,2}"]
+        file_id = 123432423
+        job = PNIDAPI.extract_pattern(file_id, patterns=patterns)
+        assert isinstance(job, ContextualizationJob)
+        assert "Queued" == job.status
+        assert "items" in job.result
+        assert "Completed" == job.status
+        assert 456 == job.job_id
+
+        n_detect_calls = 0
+        n_status_calls = 0
+        for call in mock_detect.calls:
+            if "extractpattern" in call.request.url and call.request.method == "POST":
+                n_detect_calls += 1
+                assert {"patterns": patterns, "fileId": file_id} == jsgz_load(call.request.body)
+            else:
+                n_status_calls += 1
+                assert "/456" in call.request.url
+        assert 1 == n_detect_calls
+        assert 1 == n_status_calls